+++ src/main/java/me/scoltbr/scoltEconomys/stats/EconomyDailyRow.java
@@ -1,13 +1,14 @@
 package me.scoltbr.scoltEconomys.stats;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record EconomyDailyRow(
         LocalDate day,
-        double totalCoins,
-        double totalWallet,
-        double totalBank,
+        BigDecimal totalCoins,
+        BigDecimal totalWallet,
+        BigDecimal totalBank,
         int activePlayers,
-        double top10Concentration,
+        BigDecimal top10Concentration,
         long updatedAt
 ) {}
